+++ CHANGELOG.md
@@ -31,6 +31,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Documentation for `:extensions:data-plane:data-plane-api` (#1579)
 * Allow `TypeManager` to support multiple serialization contexts (#1581)
 * Check response type of IDS multipart messages (#1695)
+* Define api error detail message schema (#1490)
 
 #### Changed